+++ backend/src/middleware/auth.middleware.js
@@ -6,7 +6,7 @@ const handleAuthHeader = async (req) => {
     if (!token) throw new Error('Token missing');
 
     const decoded = jsonwebtoken.verify(token, process.env.JWT_SECRET);
-    const user = await User.findById(decoded.id).select('_id username role');
+    const user = await User.findById(decoded.id).select('_id id username role');
     if (!user) throw new Error('User not found');
 
     return user;