+++ frontend/src/pages/Login/LoginPage.tsx
@@ -38,7 +38,6 @@ export default function LoginPage() {
             navigate(ROUTES.home);
             return;
         }
-
     }, [isChecked]);
 
     if (!isChecked) {